+++ serde-encrypt-core/src/encrypt/plain_message_public_key.rs
@@ -1,24 +1,36 @@
 //! Shared key encryption.
 
+use core::ops::DerefMut;
+
 use crate::{
     error::Error,
     key::combined_key::{ReceiverCombinedKey, SenderCombinedKey},
-    random::global_rng,
+    random::RngSingleton,
 };
 use alloc::vec::Vec;
 use chacha20poly1305::{aead::Payload, XNonce};
-use core::ops::DerefMut;
 use crypto_box::{aead::Aead, ChaChaBox};
 
 use super::encrypted_message::EncryptedMessage;
 
 /// Plain message structure serialized via serde.
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct PlainMessagePublicKey(Vec<u8>);
+pub trait PlainMessagePublicKeyCore {
+    /// RNG singleton
+    type R: RngSingleton;
 
-impl PlainMessagePublicKey {
+    /// Constructor
+    fn new(plain_message: Vec<u8>) -> Self
+    where
+        Self: Sized;
+
+    /// Raw representation
+    fn into_vec(self) -> Vec<u8>;
+
+    /// Ref to raw representation
+    fn as_slice(&self) -> &[u8];
+
     /// Encrypt into EncryptedMessage
-    pub fn encrypt(&self, combined_key: &SenderCombinedKey) -> Result<EncryptedMessage, Error> {
+    fn encrypt(&self, combined_key: &SenderCombinedKey) -> Result<EncryptedMessage, Error> {
         let nonce = Self::generate_nonce();
         let sender_box = ChaChaBox::new(
             combined_key.receiver_public_key().as_ref(),
@@ -29,7 +41,13 @@ impl PlainMessagePublicKey {
         let aad = b"".as_ref();
 
         let encrypted = sender_box
-            .encrypt(&nonce, Payload { msg: &self.0, aad })
+            .encrypt(
+                &nonce,
+                Payload {
+                    msg: self.as_slice(),
+                    aad,
+                },
+            )
             .map_err(|_| {
                 Error::encryption_error("failed to encrypt serialized data into ChaChaBox")
             })?;
@@ -38,10 +56,13 @@ impl PlainMessagePublicKey {
     }
 
     /// Decrypt from EncryptedMessage
-    pub fn decrypt(
+    fn decrypt(
         encrypted_message: &EncryptedMessage,
         combined_key: &ReceiverCombinedKey,
-    ) -> Result<Self, Error> {
+    ) -> Result<Self, Error>
+    where
+        Self: Sized,
+    {
         let receiver_box = ChaChaBox::new(
             combined_key.sender_public_key().as_ref(),
             combined_key.receiver_private_key().as_ref(),
@@ -54,23 +75,12 @@ impl PlainMessagePublicKey {
             .decrypt(nonce.into(), encrypted)
             .map_err(|_| Error::decryption_error("error on decryption of ChaChaBox"))?;
 
-        Ok(Self(serial_plain))
+        Ok(Self::new(serial_plain))
     }
 
+    /// Generate random nonce which is large enough (24-byte) to rarely conflict.
     fn generate_nonce() -> XNonce {
-        let mut rng = global_rng().lock();
+        let mut rng = Self::R::instance();
         crypto_box::generate_nonce(rng.deref_mut())
-    }
-}
-
-impl From<Vec<u8>> for PlainMessagePublicKey {
-    fn from(plain: Vec<u8>) -> Self {
-        Self(plain)
-    }
-}
-
-impl From<PlainMessagePublicKey> for Vec<u8> {
-    fn from(p: PlainMessagePublicKey) -> Self {
-        p.0
     }
 }