+++ serde-encrypt-core/src/encrypt/plain_message_shared_key.rs
@@ -2,27 +2,38 @@
 
 use core::ops::DerefMut;
 
-use crate::{error::Error, key::as_shared_key::AsSharedKey, random::global_rng};
+use super::encrypted_message::EncryptedMessage;
+use crate::random::RngSingleton;
+use crate::{error::Error, key::as_shared_key::AsSharedKey};
 use alloc::{format, vec::Vec};
 use chacha20poly1305::{XChaCha20Poly1305, XNonce};
 use crypto_box::aead::{Aead, NewAead};
 
-use super::encrypted_message::EncryptedMessage;
-
 /// Plain message structure serialized via serde.
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct PlainMessageSharedKey(Vec<u8>);
+pub trait PlainMessageSharedKeyCore {
+    /// RNG singleton
+    type R: RngSingleton;
 
-impl PlainMessageSharedKey {
+    /// Constructor
+    fn new(plain_message: Vec<u8>) -> Self
+    where
+        Self: Sized;
+
+    /// Raw representation
+    fn into_vec(self) -> Vec<u8>;
+
+    /// Ref to raw representation
+    fn as_slice(&self) -> &[u8];
+
     /// Encrypt into EncryptedMessage
-    pub fn encrypt<S>(&self, shared_key: &S) -> Result<EncryptedMessage, Error>
+    fn encrypt<S>(&self, shared_key: &S) -> Result<EncryptedMessage, Error>
     where
         S: AsSharedKey,
     {
         let nonce = Self::generate_nonce();
         let chacha = XChaCha20Poly1305::new(shared_key.to_chacha_key());
 
-        let encrypted = chacha.encrypt(&nonce, self.0.as_ref()).map_err(|e| {
+        let encrypted = chacha.encrypt(&nonce, self.as_slice()).map_err(|e| {
             Error::encryption_error(&format!(
                 "failed to encrypt serialized data by XChaCha20: {:?}",
                 e
@@ -33,8 +44,9 @@ impl PlainMessageSharedKey {
     }
 
     /// Decrypt from EncryptedMessage
-    pub fn decrypt<S>(encrypted_message: &EncryptedMessage, shared_key: &S) -> Result<Self, Error>
+    fn decrypt<S>(encrypted_message: &EncryptedMessage, shared_key: &S) -> Result<Self, Error>
     where
+        Self: Sized,
         S: AsSharedKey,
     {
         let nonce = encrypted_message.nonce();
@@ -48,23 +60,12 @@ impl PlainMessageSharedKey {
                 e
             ))
         })?;
-        Ok(Self(plain))
+        Ok(Self::new(plain))
     }
 
+    /// Generate random nonce which is large enough (24-byte) to rarely conflict.
     fn generate_nonce() -> XNonce {
-        let mut rng = global_rng().lock();
+        let mut rng = Self::R::instance();
         crypto_box::generate_nonce(rng.deref_mut())
-    }
-}
-
-impl From<Vec<u8>> for PlainMessageSharedKey {
-    fn from(plain: Vec<u8>) -> Self {
-        Self(plain)
-    }
-}
-
-impl From<PlainMessageSharedKey> for Vec<u8> {
-    fn from(p: PlainMessageSharedKey) -> Self {
-        p.0
     }
 }