+++ serde-encrypt/src/key.rs
@@ -1,0 +1,4 @@
+#![allow(missing_docs)]
+
+#[deny(missing_docs)]
+pub mod key_pair;